+++ tests/resources/sample_lambda_with_prebundle/sample_script.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+touch test_ramuda_prebundle.txt